+++ interface_stone_smart_flutter/lib/utils/types/payment_type_transaction.dart
@@ -1,0 +1,6 @@
+enum PaymentTypeTransaction {
+  credit,
+  debit,
+  voucher,
+  pix,
+}