+++ stone_smart_flutter/android/src/main/java/com/qztech/stone_smart_flutter/core/StoneSmart.java
@@ -29,9 +29,7 @@ public class StoneSmart {
 
     private static final String PAYMENT_REVERSAL = "paymentReversal";
     private static final String PAYMENT_OPTIONS = "paymentOptions";
-
-    private static final String PAYMENT_PRINTER_TRANSACTION = "paymentPrinterTransaction";
-
+    private static final String PAYMENT_PRINT_TRANSACTION = "paymentPrinterTransaction";
     private static final String PAYMENT_GET_TRANSACTION_BY_INITIATOR_TRANSACTION_KEY = "paymentGetTransactionByInitiatorTransactionKey";
 
     private static final String PAYMENT_GET_ALL_TRANSACTIONS = "paymentGetAllTransactions";
@@ -59,7 +57,7 @@ public class StoneSmart {
 
         if (call.method.equals(PAYMENT_PRINTER_BASE64)) {
             String printerParams = call.argument("printerParams");
-            this.payment.printerFromBase64(printerParams, currentContext);
+            this.payment.printFromBase64(printerParams, currentContext);
             return;
         }
 
@@ -74,14 +72,14 @@ public class StoneSmart {
         }
         
         if(call.method.equals(PAYMENT_GET_TRANSACTION_BY_INITIATOR_TRANSACTION_KEY)) {
-            String initiatorTransactionKey = call.argument("initiatorTransactionKey");
-            this.payment.getTransactionByInitiatorTransactionKey(currentContext, initiatorTransactionKey);
+            String initiatorKey = call.argument("initiatorKey");
+            this.payment.getTransactionByInitiatorTransactionKey(currentContext, initiatorKey);
             return;
         }
 
-        if(call.method.equals(PAYMENT_PRINTER_TRANSACTION)) {
-            boolean isPrinter = call.argument("isPrinter");
-            this.payment.printerCurrentTransaction(currentContext, isPrinter);
+        if(call.method.equals(PAYMENT_PRINT_TRANSACTION)) {
+            boolean printCustomerSlip = call.argument("printCustomerSlip");
+            this.payment.printerCurrentTransaction(currentContext, printCustomerSlip);
             return;
         }
 
@@ -127,21 +125,21 @@ public class StoneSmart {
 
         int parc = call.argument("installment");
         boolean withInterest = call.argument("withInterest");
-        boolean isPrinter = call.argument("isPrinter");
-        String initiatorTransactionKey = call.argument("initiatorTransactionKey");
+        boolean printCustomerSlip = call.argument("printCustomerSlip");
+        String initiatorKey = call.argument("initiatorKey");
 
         if (call.method.equals(PAYMENT_DEBIT)) {
-            this.payment.doTransaction(currentContext,amount, 2, initiatorTransactionKey, parc, withInterest, null,null, isPrinter);
+            this.payment.doTransaction(currentContext,amount, 2, initiatorKey, parc, withInterest, null,null, printCustomerSlip);
         } else if (call.method.equals(PAYMENT_PIX)) {
             String qrCodeAuthotization = call.argument("qrCodeAuthorization");
             String qrCodeProviderid = call.argument("qrCodeProviderid");
-            this.payment.doTransaction(currentContext,amount, 3, initiatorTransactionKey, parc, withInterest, qrCodeAuthotization, qrCodeProviderid, isPrinter);
+            this.payment.doTransaction(currentContext,amount, 3, initiatorKey, parc, withInterest, qrCodeAuthotization, qrCodeProviderid, printCustomerSlip);
         } else if (call.method.equals(PAYMENT_CREDIT)) {
-            this.payment.doTransaction(currentContext,amount, 1, initiatorTransactionKey, parc, withInterest,null,null, isPrinter);
+            this.payment.doTransaction(currentContext,amount, 1, initiatorKey, parc, withInterest,null,null, printCustomerSlip);
         }  else if (call.method.equals(PAYMENT_CREDIT_PARC)) {
-            this.payment.doTransaction(currentContext,amount, 1, initiatorTransactionKey, parc, withInterest,null,null, isPrinter);
+            this.payment.doTransaction(currentContext,amount, 1, initiatorKey, parc, withInterest,null,null, printCustomerSlip);
         } else if (call.method.equals(PAYMENT_VOUCHER)) {
-            this.payment.doTransaction(currentContext,amount, 4,initiatorTransactionKey, parc, withInterest, null,null, isPrinter);
+            this.payment.doTransaction(currentContext,amount, 4,initiatorKey, parc, withInterest, null,null, printCustomerSlip);
         } else {
             result.notImplemented();
         }