+++ stone_smart_flutter/example/lib/payment/payment_page.dart
@@ -33,7 +33,9 @@ class _PaymentPageState extends State<PaymentPage> {
           ),
           TextField(
             onChanged: (value) => setState(() {
-              controller.setSaleValue(moneyController.text.isNotEmpty ? double.parse(moneyController.text) : 0.0);
+              controller.setSaleValue(moneyController.text.isNotEmpty
+                  ? double.parse(moneyController.text)
+                  : 0.0);
             }),
             keyboardType: TextInputType.number,
             decoration: const InputDecoration(hintText: "Digite o valor"),
@@ -69,7 +71,9 @@ class _PaymentPageState extends State<PaymentPage> {
                         setState(() {
                           controller.clickPayment = true;
                         });
-                        StoneSmart.instance().payment.creditPayment(controller.saleValue);
+                        StoneSmart.instance()
+                            .payment
+                            .creditPayment(controller.saleValue);
                       }
                     : null,
                 child: const Text("Crédito"),
@@ -97,7 +101,9 @@ class _PaymentPageState extends State<PaymentPage> {
                         setState(() {
                           controller.clickPayment = true;
                         });
-                        StoneSmart.instance().payment.voucherPayment(controller.saleValue);
+                        StoneSmart.instance()
+                            .payment
+                            .voucherPayment(controller.saleValue);
                       }
                     : null,
                 child: const Text("Voucher"),
@@ -109,7 +115,9 @@ class _PaymentPageState extends State<PaymentPage> {
                   setState(() {
                     controller.clickPayment = true;
                   });
-                  StoneSmart.instance().payment.activePinpad(appName: 'AppDemo', stoneCode: '206192723');
+                  StoneSmart.instance()
+                      .payment
+                      .activePinpad(appName: 'AppDemo', stoneCode: '206192723');
                 },
               ),
             ],