+++ stone_smart_flutter/CHANGELOG.md
@@ -34,7 +34,7 @@
 
 ## 1.0.16
 
-- New methods `customPrinter` and `printerFromBase64` from printer.
+- New methods `customPrinter` and `printFromBase64` from printer.
 
 ## 1.0.15